+++ lib/i18n/tasks/data/tree/node.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 
 require 'i18n/tasks/data/tree/traversal'
 require 'i18n/tasks/data/tree/siblings'
@@ -71,6 +72,10 @@ module I18n::Tasks::Data::Tree
 
     def data?
       @data.present?
+    end
+
+    def reference?
+      value.is_a?(Symbol)
     end
 
     def get(key)
@@ -154,10 +159,18 @@ module I18n::Tasks::Data::Tree
                 Term::ANSIColor.dark '∅'
               else
                 [Term::ANSIColor.color(1 + level % 15, key),
-                 (": #{Term::ANSIColor.cyan(value.to_s)}" if leaf?),
+                 (": #{format_value_for_inspect(value)}" if leaf?),
                  (" #{data}" if data?)].compact.join
               end
       ['  ' * level, label, ("\n" + children.map { |c| c.inspect(level + 1) }.join("\n") if children?)].compact.join
+    end
+
+    def format_value_for_inspect(value)
+      if value.is_a?(Symbol)
+        "#{Term::ANSIColor.bold(Term::ANSIColor.yellow '⮕ ')}#{Term::ANSIColor.yellow value.to_s}"
+      else
+        Term::ANSIColor.cyan(value.to_s)
+      end
     end
 
     protected