+++ lib/i18n/tasks/reports/base.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 module I18n::Tasks::Reports
   class Base
     include I18n::Tasks::Logging
@@ -27,11 +28,9 @@ module I18n::Tasks::Reports
       "Same value as #{locale} (#{key_values.count || '∅'})"
     end
 
-    def used_title(used_tree)
-      leaves = used_tree.leaves.to_a
-      filter = used_tree.first.root.data[:key_filter]
-      used_n = leaves.map { |node| node.data[:occurrences].size }.reduce(:+).to_i
-      "#{leaves.length} key#{'s' if leaves.size != 1}#{" matching '#{filter}'" if filter}#{" (#{used_n} usage#{'s' if used_n != 1})" if used_n > 0}"
+    def used_title(keys_nodes, filter)
+      used_n = keys_nodes.map { |_k, node| node.data[:occurrences].size }.reduce(:+).to_i
+      "#{keys_nodes.size} key#{'s' if keys_nodes.size != 1}#{" matching '#{filter}'" if filter}#{" (#{used_n} usage#{'s' if used_n != 1})" if used_n > 0}"
     end
 
     # Sort keys by their attributes in order
@@ -58,6 +57,12 @@ module I18n::Tasks::Reports
       else
         locale.tr '+', ' '
       end
+    end
+
+    def collapse_missing_tree!(forest)
+      forest = task.collapse_plural_nodes!(forest)
+      forest = task.collapse_same_key_in_locales!(forest) { |node| node.data[:type] == :missing_used }
+      forest
     end
   end
 end