+++ lib/i18n/tasks/scanners/ruby_ast_scanner.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 require 'i18n/tasks/scanners/file_scanner'
 require 'i18n/tasks/scanners/relative_keys'
 require 'i18n/tasks/scanners/ruby_ast_call_finder'
@@ -75,7 +76,7 @@ module I18n::Tasks::Scanners
                    else
                      key
                    end
-        [full_key, range_to_occurrence(location.expression, default_arg: default_arg)]
+        [full_key, range_to_occurrence(key, location.expression, default_arg: default_arg)]
       end
     end
 
@@ -163,16 +164,18 @@ module I18n::Tasks::Scanners
       /controllers|mailers/.match(path)
     end
 
+    # @param raw_key [String]
     # @param range [Parser::Source::Range]
     # @param default_arg [String, nil]
     # @return [Results::Occurrence]
-    def range_to_occurrence(range, default_arg: nil)
+    def range_to_occurrence(raw_key, range, default_arg: nil)
       Results::Occurrence.new(
           path:        range.source_buffer.name,
           pos:         range.begin_pos,
           line_num:    range.line,
           line_pos:    range.column,
           line:        range.source_line,
+          raw_key:     raw_key,
           default_arg: default_arg)
     end