+++ lib/i18n/tasks/unused_keys.rb
@@ -1,3 +1,4 @@
+# frozen_string_literal: true
 require 'set'
 
 module I18n
@@ -11,10 +12,11 @@ module I18n
       # @param [String] locale
       # @param [Boolean] strict if true, do not match dynamic keys
       def unused_tree(locale: base_locale, strict: nil)
+        used_key_names = used_tree(strict: true).key_names
         collapse_plural_nodes! data[locale].select_keys { |key, _node|
           !ignore_key?(key, :unused) &&
               (strict || !used_in_expr?(key)) &&
-              !used_key?(depluralize_key(key, locale))
+              !used_key_names.include?(depluralize_key(key, locale))
         }
       end
     end